# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PRM Marketplace API",
    "description" : "Backend API for the student/resident marketplace",
    "contact" : {
      "name" : "PRM Marketplace Team"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/vendor-profiles/{id}" : {
      "get" : {
        "tags" : [ "vendor-profile-controller" ],
        "operationId" : "read",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VendorProfile"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "vendor-profile-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VendorProfile"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VendorProfile"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vendor-profiles" : {
      "get" : {
        "tags" : [ "vendor-profile-controller" ],
        "operationId" : "getAll_2",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/VendorProfile"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "vendor-profile-controller" ],
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VendorProfile"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VendorProfile"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/vendor/{vendorId}" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getByVendor",
        "parameters" : [ {
          "name" : "vendorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "VendorProfile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "readOnly" : true
          },
          "businessName" : {
            "type" : "string"
          },
          "registrationNo" : {
            "type" : "string"
          },
          "verified" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "ratingAvg" : {
            "type" : "number",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
...
      "Product" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "stockQuantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "category" : {
            "type" : "string"
          },
          "imageUrl" : {
            "type" : "string"
          },
          "condition" : {
            "type" : "string",
            "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
          },
          "city" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "vendor" : {
            "$ref" : "#/components/schemas/VendorProfile",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}